바텀바 생성
