--- app/src/main/java/com/example/gravit/AppNavigation.kt
@@ -1,17 +1,0 @@
-package com.example.gravit
-import androidx.compose.runtime.Composable
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import com.example.gravit.login.LoginScreen
-import com.example.gravit.splash.SplashScreen
-
-@Composable
-fun AppNavigation() {
-    val navController = rememberNavController()
-
-    NavHost(navController = navController, startDestination = "splash") {
-        composable("splash") { SplashScreen(navController) }
-        composable("login_choice") { LoginScreen(navController) }
-    }
-}
