+++ app/src/main/java/com/example/gravit/navigation/AppNavigation.kt
@@ -1,0 +1,40 @@
+package com.example.gravit
+import androidx.compose.runtime.Composable
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.example.gravit.login.LoginScreen
+import com.example.gravit.splash.SplashScreen
+
+@Composable
+fun AppNavigation() {
+    val navController = rememberNavController()
+
+    NavHost(navController = navController, startDestination = "splash") {
+        composable("splash") { SplashScreen(navController) }
+        composable("login_choice") { LoginScreen(navController) }
+        composable("main") { MainScreen() }
+    }
+}
+
+@Composable
+fun MainScreen() {
+    val bottomNavController = rememberNavController()
+
+    Scaffold(
+        bottomBar = {
+            BottomNavigationBar(bottomNavController)
+        }
+    ) { innerPadding ->
+        NavHost(
+            navController = bottomNavController,
+            startDestination = "home",
+            modifier = Modifier.padding(innerPadding)
+        ) {
+            composable("home") { HomeScreen() }
+            composable("study") { StudyScreen() }
+            composable("league") { LeagueScreen() }
+            composable("user") { UserScreen() }
+        }
+    }
+}
